% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_fractions.R
\name{active_fractions}
\alias{active_fractions}
\title{active_fractions}
\usage{
active_fractions(
  data_reporter,
  model_type,
  andor = "OR",
  variable,
  type_ctrl,
  thresh_ctrl,
  final_msg = TRUE
)
}
\arguments{
\item{data_reporter}{The reporter dataframe (data_reporter())}

\item{model_type}{Character. A model type that will be used to filter the data.}

\item{andor}{Character. If "AND" the function returns only the fractions contained in every model type, otherwhise every fraction. By default "OR".}

\item{variable}{Character. In which variable search for active fractions. e.g. Concentration.average or GFP.average...}

\item{type_ctrl}{Character. The control to use for the comparison. One of "BETA_GLUCO_C20_1_S" (trem2) or "MeOH" (seap).}

\item{thresh_ctrl}{Numeric. Percentage for the threshold of the CTRL/CTRL+. e.g. 50 is half of the CTRL, 200 is 2 times the CTRL.}

\item{final_msg}{Logical. If FALSE the final message of completed is not sent.}
}
\description{
\code{active_fractions} Search for active fractions.
}
