% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_targetfile.R
\name{check_targetfile}
\alias{check_targetfile}
\title{check_targetfile}
\usage{
check_targetfile(target, explist, check_back = TRUE)
}
\arguments{
\item{target}{The target file (dataframe).}

\item{explist}{The experiment list file.}

\item{check_back}{Boolean. Set to TRUE if you want to check the existence of the BACKGROUND. Default to TRUE (cyto and seap). Set to FALSE for D1 and TREM2.}
}
\description{
\code{check_targetfile} Run a check on the target file. Returns a target file without the faulty experiments.
}
