% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_check_table.R
\name{make_check_table}
\alias{make_check_table}
\title{make_check_table}
\usage{
make_check_table(dt)
}
\arguments{
\item{dt}{A list containing all database.}
}
\description{
\code{make_check_table} create a table with the intersection of Product_Family with yes or no
}
