% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enriched_fractions.R
\name{productive_fractions}
\alias{productive_fractions}
\title{enriched_fractions}
\usage{
productive_fractions(data_reporter, model_type, times_ctrl = 2.5)
}
\arguments{
\item{model_type}{Character. A model type that will be used to filter the SEAP data.}

\item{data_repo}{The reporter dataframe (data_reporter())}

\item{repo_type}{Character. The type of the reporter data ("SEAP" or "TREM2".}

\item{prod_trem}{Dataframe. The output from 4 or 5 query of TREM2.}
}
\description{
\code{enriched_fractions} Function for a query in Reporter.
}
