% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryD1.R
\name{queryD1}
\alias{queryD1}
\title{queryD1}
\usage{
queryD1(
  data_D1,
  MFI = c("CD80", "CD40", "MHC-II"),
  operation = "greater than",
  thresh = 2.5,
  andor = "OR"
)
}
\arguments{
\item{data_D1}{The D1 dataframe (data_D1())}

\item{MFI}{Character. The MFI column where to search (e.g. "CD80").}

\item{operation}{Operation to perform in comparing with the control (for now one between "greater than" or "greater than or equal").}

\item{thresh}{Numeric. Threshold for how many times the control. (e.g. 2 means search for MFI greater than 2 times CTRL)}

\item{andor}{Character. If muliple MFI the operation to be performed ("AND" or "OR"). "AND" returns only the results in common with all the MFIs.}
}
\description{
\code{queryD1} Function for a query in D1
}
