% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cytoxicity.R
\name{read_cytoxicity}
\alias{read_cytoxicity}
\title{read_cytoxicity}
\usage{
read_cytoxicity(
  ifile,
  path = getwd(),
  instrument = "EZ_READ_2000",
  scan = "Single",
  sample.anno = NULL,
  sep = ",",
  filter.na = "Product"
)
}
\arguments{
\item{X}{X}
}
\description{
\code{read_cytoxicity} draw multiple break line (br())
}
