% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_D1.R
\name{read_D1}
\alias{read_D1}
\title{read_D1}
\usage{
read_D1(file_explist, file_target, filter.na = "Product")
}
\arguments{
\item{file_explist}{The row of the experiment list of that sample (i.e. the experiment list filtered by EXPID)}

\item{file_target}{The row of the target file of that sample (i.e. the target file filtered by EXPID)}

\item{filter.na}{A column where the NA filtering is performed. By default is Product.}
}
\description{
\code{read_D1} Read D1 data
}
