% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_cytoxicity.R
\name{summarise_cytoxicity}
\alias{summarise_cytoxicity}
\title{eval_cytoxicity}
\usage{
summarise_cytoxicity(X, group, method = "mean")
}
\arguments{
\item{X}{X}
}
\description{
\code{read_cytoxicity} draw multiple break line (br())
}
