% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ADViSE_fresh.R
\docType{data}
\name{theme_ADViSE_fresh}
\alias{theme_ADViSE_fresh}
\title{theme_ADViSE_fresh}
\format{
An object of class \code{css} (inherits from \code{html}, \code{character}) of length 1.
}
\usage{
theme_ADViSE_fresh
}
\value{
Object produced by fresh
}
\description{
Poor Man's Flatly theme for a shinydashboard application edited
}
\seealso{
\code{\link{fresh}}
}
\keyword{datasets}
