% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_database.R
\name{update_database}
\alias{update_database}
\title{update_database}
\usage{
update_database(old_data, new_data = NULL)
}
\arguments{
\item{old_data}{List. The internal database data in the form of list(myprocesseddata,mydataset,exp_list).}

\item{new_data}{List. The new data coming from the mod_update_data module in the form of list(myprocesseddata,mydataset,exp_list).}
}
\description{
\code{update_database} Function to update the database.
}
