% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_cytoxicity.R
\name{eval_cytoxicity}
\alias{eval_cytoxicity}
\title{eval_cytoxicity}
\usage{
eval_cytoxicity(X, ref = "Product", background = "BACKGROUND", CTRL = "CTRL")
}
\arguments{
\item{X}{X}
}
\description{
\code{eval_cytoxicity} draw multiple break line (br())
}
