% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make a Heatmap using ComplexHeatmap package.}
\usage{
make_heatmap(
  data,
  add_rowannot = "Model_Family",
  add_colannot = NULL,
  title = "CBC150",
  order_data = TRUE,
  row_dend = TRUE,
  row_nclust = 2,
  col_dend = FALSE,
  col_nclust = 3,
  dist_method = "euclidean",
  clust_method = "ward.D2",
  unit_legend = "\% toxicty",
  col_label_size = 8,
  color_scale = circlize::colorRamp2(c(0, 100), c("white", "blue")),
  add_values = FALSE,
  thresh_values = 0,
  typeeval_heat = "Cytotoxicity.average"
)
}
\arguments{
\item{data}{A summarizedexperiment object. Tested only on sumexp averaged}

\item{add_rowannot}{A column used for the left (row) annotation (e.g. "Product_Batch")}

\item{add_colannot}{A column used for the bottom (column) annotation (e.g. "Class")}

\item{title}{A label for the column (i.e. "Lipids").}

\item{order_data}{Boolean. Set to TRUE if you want to order.}

\item{row_dend}{Boolean. Render dendogram on rows (TRUE or FALSE).}

\item{row_nclust}{Numeric. Number of cluster in the row dendrogram.}

\item{col_dend}{Boolean. Render dendogram on columns (TRUE or FALSE).}

\item{col_nclust}{Numeric. Number of cluster in the column dendrogram.}

\item{dist_method}{Choose a distance method (e.g "euclidean"). See ?stats::dist().}

\item{clust_method}{Choose a clustering method (e.g "ward.D"). See ?stats::hclust().}

\item{unit_legend}{An unit label for the legend. If data are scaled will be "Z-score" otherwise will be the input.}

\item{col_label_size}{Size of the column labels. By default col_label_size = 10.}

\item{color_scale}{Color scale for the cells. By default is circlize::colorRamp2(c(0, 100), c("white", "blue")).}

\item{add_values}{Boolean. Set to TRUE if you want to show cell numbers.}

\item{thresh_values}{Numeric. The threshold for the cell number.}

\item{typeeval_heat}{String. The column used for the heatmap. By default "Cytotoxicity.average".}
}
\description{
This function render a Heatmap using ComplexHeatmap package.
}
