% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_filtering.R
\name{operation_filtering}
\alias{operation_filtering}
\title{operation_filtering}
\usage{
operation_filtering(data, column, operator, value)
}
\arguments{
\item{data}{The summarized data}

\item{column}{Character. The column where to apply the filtering (e.g. "Cytotoxicity.average")}

\item{operator}{Character. The operator to be used. A choice beetween "max", "min", "greater than", "greater than or equal", "equal", "less than or equal" or "less than".}

\item{value}{Numeric. Value used to compare the data.}
}
\description{
\code{operation_filtering} Function for the query
}
